tests/api-resources/evaluate/question-answering.test.ts
// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.

import LlamaStackClient from 'llama-stack-client';
import { Response } from 'node-fetch';

const client = new LlamaStackClient({ baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010' });

describe('resource questionAnswering', () => {
  test('create: only required params', async () => {
    const responsePromise = client.evaluate.questionAnswering.create({ metrics: ['em', 'f1'] });
    const rawResponse = await responsePromise.asResponse();
    expect(rawResponse).toBeInstanceOf(Response);
    const response = await responsePromise;
    expect(response).not.toBeInstanceOf(Response);
    const dataAndResponse = await responsePromise.withResponse();
    expect(dataAndResponse.data).toBe(response);
    expect(dataAndResponse.response).toBe(rawResponse);
  });

  test('create: required and optional params', async () => {
    const response = await client.evaluate.questionAnswering.create({
      metrics: ['em', 'f1'],
      'X-LlamaStack-ProviderData': 'X-LlamaStack-ProviderData',
    });
  });
});
